+++ src/External/Restaurant.Infrastructure/Repositories/Catalog/IngredientCategoryRepository.cs
@@ -16,7 +16,7 @@ namespace Restaurant.Infrastructure.Repositories.Catalog
             return await _context.IngredientCategories.AnyAsync(x => EF.Functions.ILike(x.Name, name), cancellationToken);
         }
 
-        public async Task<IngredientCategory?> FindByIdAsync(int id, CancellationToken cancellationToken = default)
+        public async Task<IngredientCategory?> FindByIdAsync(long id, CancellationToken cancellationToken = default)
         {
             return await _context.IngredientCategories.FirstOrDefaultAsync(x => x.Id == id, cancellationToken);
         }
